add completer of interactive mode
