+++ src/test/java/com/github/imas/rdflint/RdfLintTest.java
@@ -1,21 +1,12 @@
 package com.github.imas.rdflint;
 
 import static junit.framework.TestCase.assertEquals;
-import static junit.framework.TestCase.assertFalse;
-import static junit.framework.TestCase.assertTrue;
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.junit.Assert.assertNotEquals;
 import static org.junit.Assert.assertNull;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 import com.github.imas.rdflint.config.RdfLintParameters;
-import java.io.ByteArrayOutputStream;
-import java.io.File;
-import java.io.IOException;
 import org.apache.commons.cli.CommandLine;
-import org.apache.jena.rdf.model.Model;
-import org.hamcrest.CoreMatchers;
 import org.junit.Test;
 
 public class RdfLintTest {
@@ -111,51 +102,5 @@ public class RdfLintTest {
     assertNull("getOriginDir", params.getOriginDir());
   }
 
-  @Test
-  public void interactiveCommand() throws Exception {
-    RdfLint lint = new RdfLint();
-    RdfLintParameters params = lint.loadConfig(getParentPath("config_ok/rdflint-config.yml"));
-    params.setTargetDir(getParentPath("config_ok/"));
-    String parentPath = new File(params.getTargetDir()).getCanonicalPath();
-    Model m = lint.loadRdfSet(params, params.getTargetDir());
-
-    // exit
-    ByteArrayOutputStream byteOut = new ByteArrayOutputStream();
-    assertFalse(":exit", lint.interactiveCommand(byteOut, ":exit",
-        params, params.getTargetDir(), parentPath, m));
-    assertEquals("exit out", 0, byteOut.toString("UTF-8").length());
-
-    assertEquals("reload out",
-        0, interactiveCommandCall(":reload", lint, params, parentPath, m).length());
-    assertEquals("check out",
-        0, interactiveCommandCall(":check", lint, params, parentPath, m).length());
-    assertThat("help out",
-        interactiveCommandCall(":help", lint, params, parentPath, m),
-        CoreMatchers.containsString("help"));
-    assertNotEquals("unknown out",
-        0, interactiveCommandCall(":unknown", lint, params, parentPath, m).length());
-
-    assertThat("select out",
-        interactiveCommandCall("select ?s ?p ?o where {?s ?p ?o}", lint, params, parentPath, m),
-        CoreMatchers.containsString("familyName"));
-    assertThat("describe out",
-        interactiveCommandCall("describe <https://sparql.crssnky.xyz/imasrdf/something>",
-            lint, params, parentPath, m),
-        CoreMatchers.containsString("\"familyName\"@ja"));
-    assertThat("ask out",
-        interactiveCommandCall("ask {<https://sparql.crssnky.xyz/imasrdf/something> ?p ?o}", lint,
-            params, parentPath, m),
-        CoreMatchers.containsString("true"));
-  }
-
-  private String interactiveCommandCall(
-      String line, RdfLint lint, RdfLintParameters params, String parentPath, Model m)
-      throws IOException {
-    ByteArrayOutputStream byteOut = new ByteArrayOutputStream();
-    boolean rtn = lint
-        .interactiveCommand(byteOut, line, params, params.getTargetDir(), parentPath, m);
-    assertTrue(line, rtn);
-    return byteOut.toString("UTF-8");
-  }
 
 }
